Added saving metrics to CSV --metricsfile --metricsfreq

+++ xtransmit/metrics_jitter.hpp
@@ -8,25 +8,25 @@ namespace metrics
 
 class jitter_trace
 {
-    typedef std::chrono::steady_clock::time_point time_point;
-    typedef std::chrono::steady_clock::duration duration;
+	typedef std::chrono::steady_clock::time_point time_point;
+	typedef std::chrono::steady_clock::duration duration;
 
 public:
-    jitter_trace()
-    {}
+	jitter_trace()
+	{}
 
 public:
 	/// Submit new sample for jitter update.
 	/// @param [in] sample_time  the timestamp of the sample
 	/// @param [in] current_time current time to compare the timestamp with
-    void new_sample(const time_point& sample_time, const time_point& current_time);
+	void new_sample(const time_point& sample_time, const time_point& current_time);
 
 	/// Get curent jitter value.
-    uint64_t jitter() const { return m_jitter; }
+	uint64_t jitter() const { return (uint64_t) m_jitter; }
 
 private:
-    duration m_prev_delay = duration::zero();
-    uint64_t m_jitter = 0;
+	duration m_prev_delay = duration::zero();
+	double m_jitter = 0.0;
 };
